:truck: Move ReproducibilityRate to certif contexte
