+++ api/src/certification/shared/domain/models/ReproducibilityRate.js
