+++ api/src/shared/domain/constants.js
@@ -1,3 +1,4 @@
+import { MINIMUM_REPRODUCIBILITY_RATE_TO_BE_CERTIFIED } from '../../certification/shared/domain/constants.js';
 import { config } from '../config.js';
 
 const LEVENSHTEIN_DISTANCE_MAX_RATE = 0.25;
@@ -93,7 +94,6 @@ const MINIMUM_DELAY_IN_DAYS_BEFORE_RETRYING = config.features.dayBeforeRetrying;
 
 const MINIMUM_CERTIFIABLE_COMPETENCES_FOR_CERTIFIABILITY = 5;
 const MINIMUM_COMPETENCE_LEVEL_FOR_CERTIFIABILITY = 1;
-const MINIMUM_REPRODUCIBILITY_RATE_TO_BE_CERTIFIED = 50;
 const MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED = 80;
 
 const V3_REPRODUCIBILITY_RATE = 100;
@@ -198,7 +198,6 @@ export {
   MINIMUM_DELAY_IN_DAYS_BEFORE_IMPROVING,
   MINIMUM_DELAY_IN_DAYS_BEFORE_RETRYING,
   MINIMUM_DELAY_IN_DAYS_FOR_RESET,
-  MINIMUM_REPRODUCIBILITY_RATE_TO_BE_CERTIFIED,
   MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED,
   OIDC_ERRORS,
   ORGANIZATION_FEATURE,
