+++ api/tests/tooling/domain-builder/factory/build-reproducibility-rate.js
@@ -1,4 +1,4 @@
-import { ReproducibilityRate } from '../../../../src/shared/domain/models/ReproducibilityRate.js';
+import { ReproducibilityRate } from '../../../../src/certification/shared/domain/models/ReproducibilityRate.js';
 
 const buildReproducibilityRate = function ({ value = 10 } = {}) {
   return new ReproducibilityRate(value);
